Work on fastlane screenshot
