+++ App/dfuUITests/SnapshotHelper.swift
@@ -22,7 +22,7 @@ func setupSnapshot(_ app: XCUIApplication, waitForAnimations: Bool = true) {
     Snapshot.setupSnapshot(app, waitForAnimations: waitForAnimations)
 }
 
-func snapshot(_ name: String, waitForLoadingsetUpIndicator: Bool) {
+func snapshot(_ name: String, waitForLoadingIndicator: Bool) {
     if waitForLoadingIndicator {
         Snapshot.snapshot(name)
     } else {
