+++ App/fastlane/Deliverfile.swift
@@ -1,0 +1,13 @@
+// The Deliverfile allows you to store various App Store Connect metadata
+// For more information, check out the docs
+// https://docs.fastlane.tools/actions/deliver/
+
+// In general, you can use the options available
+// fastlane deliver --help
+
+// Remove the // in front of the line to enable the option
+
+class Deliverfile: DeliverfileProtocol {
+  //var username: String { return "" }
+  //var appIdentifier: String? { return "" }
+}
