+++ App/fastlane/Preview.html
@@ -1,0 +1,232 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>deliver - Device Firmware Update</title>
+        <meta http-equiv="content-type" content="text/html; charset=UTF-8">
+        <style>
+          .app-name {
+              font-size: 42px;
+              font-family: 'Helvetica Neue', HelveticaNeue, Helvetica Neue;
+              font-weight: 300;
+              margin-top: 22px;
+              margin-left: 25px;
+              margin-right: 25px;
+          }
+
+          .app-subtitle {
+              font-size: 26px;
+              font-family: 'Helvetica Neue', HelveticaNeue, Helvetica Neue;
+              font-weight: 300;
+              margin-left: 25px;
+              margin-right: 25px;
+          }
+
+          .app-urls {
+              margin-left: 25px;
+              font-size: 14px;
+              margin-top: 4px;
+              margin-right: 25px;
+          }
+
+          body {
+              font-family: 'Helvetica Neue', HelveticaNeue;
+          }
+
+          hr {
+            margin-top: 40px;
+            margin-bottom: 40px;
+          }
+
+          .app-url-descr {
+              height:22px;
+              display: inline-block;
+              font-weight: 500;
+          }
+
+          .app-url {
+              color: #0056ba;
+              font-weight: 300;
+              margin-right: 10px;
+          }
+
+          .app-keyword {
+              margin-left: 25px;
+              margin-right: 25px;
+              margin-top: 22px;
+          }
+
+          .cat-headline {
+              font-weight: 700;
+              margin-bottom: 5px;
+          }
+
+
+
+          .app-description {
+              margin-left: 25px;
+              margin-right: 25px;
+              margin-top: 22px;
+          }
+
+          .app-description-text {
+              font-weight: 300;
+          }
+
+          .app-changelog {
+              margin-left: 25px;
+              margin-right: 25px;
+              margin-top: 22px;
+              font-weight: 300;
+          }
+
+          .app-screenshots {
+              margin-left: 15px;
+              margin-right: 15px;
+              margin-top: 22px;
+          }
+
+          .app-keyword-list {
+              font-weight: 300;
+          }
+
+          .app-changelog-list {
+              list-style-type: square;
+
+              font-weight: 300;
+          }
+
+          ul {
+              margin-top: 6px;
+          }
+
+          .app-screenshot-row {
+              width: 100%;
+              margin-top: 10px;
+              margin-bottom: 10px;
+          }
+
+          .app-screenshot {
+              width: calc(20% - 30px);
+              margin-left: 10px;
+              margin-right: 10px;
+          }
+
+          #app-screenshots .cat-headline {
+              margin-left: 10px;
+              margin-right: 10px;
+          }
+          .app-icons {
+            overflow: hidden;
+          }
+          .app-icons img {
+            width: 150px;
+          }
+          .app-icons .app-icon {
+              float: left;
+              margin-right: 20px;
+          }
+
+          .app-minor-information {
+            margin-left: 15px;
+            margin-right: 15px;
+            margin-top: 22px;
+          }
+
+          .app-minor-information-key {
+            font-weight: 700;
+          }
+        </style>
+    </head>
+
+    <body>
+      <div class="app-icons">
+
+        
+        
+      </div>
+      
+        
+          <div class="app-name">
+            en-US
+          </div>
+        
+
+        
+
+        <div class="app-urls">
+          
+          
+            
+            
+          
+            
+            
+          
+        </div>
+
+        
+
+        
+
+        
+
+        
+
+        <div class="app-screenshots">
+          <div class="cat-headline">Screenshots</div>
+
+          
+            
+            
+
+            
+              
+              
+              <h4>iPhone 6 Plus</h4>
+              <div class="app-screenshot-row">
+
+              
+                <a href="screenshots/en-US/iPhone%208%20Plus-01Devices.png" target="_blank"><img class="app-screenshot" src="screenshots/en-US/iPhone%208%20Plus-01Devices.png" title="Screenshot #0 for en-US"></a>
+              
+              </div>
+            
+              
+              
+              <h4>iPhone XS Max</h4>
+              <div class="app-screenshot-row">
+
+              
+                <a href="screenshots/en-US/iPhone%2013%20Pro%20Max-01Devices.png" target="_blank"><img class="app-screenshot" src="screenshots/en-US/iPhone%2013%20Pro%20Max-01Devices.png" title="Screenshot #0 for en-US"></a>
+              
+              </div>
+            
+              
+              
+              <h4>iPad Pro</h4>
+              <div class="app-screenshot-row">
+
+              
+                <a href="screenshots/en-US/iPad%20Pro%20(12.9-inch)%20(2nd%20generation)-01Devices.png" target="_blank"><img class="app-screenshot" src="screenshots/en-US/iPad%20Pro%20(12.9-inch)%20(2nd%20generation)-01Devices.png" title="Screenshot #0 for en-US"></a>
+              
+              </div>
+            
+              
+              
+              <h4>iPad Pro (12.9-inch) (3rd generation)</h4>
+              <div class="app-screenshot-row">
+
+              
+                <a href="screenshots/en-US/iPad%20Pro%20(12.9-inch)%20(5th%20generation)-01Devices.png" target="_blank"><img class="app-screenshot" src="screenshots/en-US/iPad%20Pro%20(12.9-inch)%20(5th%20generation)-01Devices.png" title="Screenshot #0 for en-US"></a>
+              
+              </div>
+            
+          
+          </div>
+
+        <hr />
+      
+      
+
+      
+    </body>
+</html>
