+++ App/fastlane/metadata/copyright.txt
@@ -1,0 +1,1 @@
+
