+++ App/fastlane/metadata/en-US/name.txt
@@ -1,0 +1,1 @@
+Device Firmware Update
