+++ App/fastlane/metadata/en-US/promotional_text.txt
@@ -1,0 +1,1 @@
+
